+++ esp_video/examples/common_components/example_video_common/include/boards/esp32-p4-function-ev-board-v1.4/example_video_common_board.h
@@ -1,0 +1,35 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: ESPRESSIF MIT
+ */
+
+#pragma once
+
+#include "sdkconfig.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * @brief ESP32-P4-Function-EV-Board V1.4 configuration
+ */
+#if CONFIG_EXAMPLE_ENABLE_MIPI_CSI_CAM_SENSOR
+/**
+ * @brief MIPI-CSI camera sensor configuration
+ */
+#define EXAMPLE_MIPI_CSI_SCCB_I2C_SCL_PIN               8
+#define EXAMPLE_MIPI_CSI_SCCB_I2C_SDA_PIN               7
+#define EXAMPLE_MIPI_CSI_CAM_SENSOR_RESET_PIN           -1
+#define EXAMPLE_MIPI_CSI_CAM_SENSOR_PWDN_PIN            -1
+#define EXAMPLE_MIPI_CSI_XCLK_PIN                       -1
+#endif /* CONFIG_EXAMPLE_ENABLE_MIPI_CSI_CAM_SENSOR */
+
+#if CONFIG_EXAMPLE_ENABLE_DVP_CAM_SENSOR
+#error "DVP interface camera sensor is not supported on ESP32-P4-Function-EV-Board V1.4 by default"
+#endif /* CONFIG_EXAMPLE_ENABLE_DVP_CAM_SENSOR */
+
+#ifdef __cplusplus
+}
+#endif